moves move_to_pickup to functions
