--- docs/api/aim.digifeeds.move_to_pickup.rst
@@ -1,7 +1,0 @@
-aim.digifeeds.move\_to\_pickup module
-=====================================
-
-.. automodule:: aim.digifeeds.move_to_pickup
-   :members:
-   :undoc-members:
-   :show-inheritance:
