+++ docs/api/aim.digifeeds.rst
@@ -25,4 +25,3 @@ Submodules
    aim.digifeeds.db_client
    aim.digifeeds.functions
    aim.digifeeds.item
-   aim.digifeeds.move_to_pickup
